Implement use of a config.ini file
Ability to set to a default value.  Ability to set to a - quick setting - for a specific output format bypassing the need to use manual settings. Validation of ini file when read from disk in the event of errors when edited manually.  When found corrupt give option for new default file or end program and edit file.  Still some work to do moving defaults into objects for easier extension as new options added and gather user settings from command line when they do not want to use defaults.