Implement support for links to other pages in local and other note books.
Links are created where link text matches a page name.
If there are duplicate page names a link to each of the duplicated pages will be created as it is not possible to tell which was the original target.
Renamed links can not be made to work unless the first link on the page to another page is valid, then renamed links will use this link.
Recycle bin is excluded from searches for pages to link to.