Simplified command line arguments.
Command line arguments simplified to options for using gui (which may not be written), or command line interface ,or ini file.

There are too many settings to use the command line to configure all the conversion options and the ini file is there for these anyway.

Fixed an issue with timers that would print time when running in silent mode (changed form decorator to context manager mode)