Tidy up, removed time in file name except for nsx, other small fixes
Fixed a spelling, removed 'any' option from metadata front matter format as was not needed.

Removed file in file name option as only known source of a time is nsx files, too many possibilities of time formats in other meta data to result in a valid entry.

Changed logic for prevention of selecting html to html conversion - had exited if user selected such a pointless conversion now removed html as export option if input is html

Fixed html and markdown parsing of meta data so when no schema provided all tags are used.

Removed redundant questions from interactive command line interface for tag prefix when using a front matter format

Fix to config_data validation, some entries are allowed to be be blank but was testing and raising error when finding a blank entry.  Fixed so blank entries are allowed