Create global value for DATA_DIR, with the DATA_DIR value being used as the root of all config.ini, logs and input and output files.  Tidies up things and helps with docker image volumes for consistency of where to find data.
Change default for source in arg_parsing.py was using cwd but this affected logic in setting source in config_settings as if it is blank we really do want blank as the setter for source in config_settings will generate the correct working directory.

Added the new DATA_DIR to the path used by source setter in config_settings

Use new DATA_DIR when creating folders for nsx data to be saved to in sn_notebook, sn_note_writer and sn_attachment_writer

Correct message if using silent mode when parsing command line arguments as mentions -g option no longer used.

Fix typo in docstring in helper_functions.py