Change silent mode approach and changes to handling working directory and source directory
Silent mode - now uses config.py to store a global silent setting - simplified the code in several places and allowed handling earlier in the process as when swapping config settings at start up there were ties when it was not set correctly.  Also simplified logic in notes_converter NotesConverter evaluate_command_line_arguments() and in arg_parsing by making -s mean silent AND useini file

Working directory - is asked for once and stored in configuration_settings.  This reduced the number of requests to calculate what it was to 2.

Cleaned up use of source path in NotesConverter and in source setter of ConversionSettings.

Removed some no longer used from NotesConverter