Changes to include pandoc in distributions
Add test for frozen environment and if so use a path to deployment package folder for pandoc.  Cold not gt pyinstaller to include the pandoc file automatically so they are `manually` added to the deployment folders.

Add exception handling for file not found thrown by subprocess in pandoc_converter.py