Tests for notes_converter.py, small refactor in yanom.py and notes_converter.py
Tests for notes converter written.
Some refactoring moving initialisation of ConfigData and parsing of the config.ini to yanom.py and merged fetching the list of nsx files to the same method as the one used to generate .md and .html file lists.

When testing over Timers in the code for NoteConverter identified a way of improving tests for timer.py and updated those tests to make them less fragile.